requirements.txt
dataclasses==0.5
Django==2.0.5
django-bootstrap3==10.0.1
djangorestframework==3.8.2
djongo==1.2.27
Pillow==5.1.0
pymongo==3.6.1
pytz==2018.4
reportlab==3.4.0
simplejson==3.15.0
sqlparse==0.2.4
xlrd==1.1.0
